current book and login check

+++ process_user_remove_book.php
@@ -7,9 +7,13 @@
 	// Connect to Database
 	include('access_database.php');
 
-	$delete = "DELETE FROM usersBooks WHERE user='".$username."' AND book='".$isbn."'";
-	$addStmt = $db->query($delete);
+	if (isset($_POST['remove'])) {
+		$delete = "DELETE FROM usersBooks WHERE user='".$username."' AND book='".$isbn."'";
+		$db->query($delete);
+	} else {
+		$update = "UPDATE users SET current_book='".$isbn."' WHERE username='".$username."'";
+		$db->query($update);
+	}
 
 	header('Location: user_homepage.php');
-
 ?>
